+++ src/App/Fossa/Test.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BlockArguments #-}
+{-# LANGUAGE OverloadedRecordDot #-}
 
 module App.Fossa.Test (
   testSubCommand,
@@ -15,7 +16,7 @@ import App.Fossa.Config.Test (
   TestOutputFormat (TestOutputJson, TestOutputPretty),
  )
 import App.Fossa.Config.Test qualified as Config
-import App.Fossa.PreflightChecks (preflightChecks)
+import App.Fossa.PreflightChecks (PreflightCommandChecks (TestChecks), preflightChecks)
 import App.Fossa.Subcommand (SubCommand)
 import App.Types (
   ProjectRevision (projectName, projectRevision),
@@ -53,7 +54,7 @@ testMain ::
   TestConfig ->
   m ()
 testMain config = do
-  _ <- ignoreDebug $ runFossaApiClient (Config.apiOpts config) preflightChecks
+  _ <- ignoreDebug $ runFossaApiClient (Config.apiOpts config) $ preflightChecks TestChecks
 
   runStickyLogger SevInfo
     . ignoreDebug -- Ignore the debug effect because we don't generate a bundle.
@@ -72,10 +73,10 @@ testMain config = do
         Nothing -> logInfo ""
         Just (DiffRevision rev) -> do
           logInfo ("diffing against revision: `" <> pretty rev <> "`")
-          logSticky $ "[ Waiting for build completion of " <> rev <> "... ]"
+          logSticky $ "[ Checking build completion for " <> rev <> "... ]"
           waitForScanCompletion revision{projectRevision = rev} cancelFlag
 
-      logSticky "[ Waiting for build completion... ]"
+      logSticky $ "[ Checking build completion for " <> revision.projectRevision <> "... ]"
       waitForScanCompletion revision cancelFlag
 
       logSticky "[ Waiting for issue scan completion... ]"