+++ Changelog.md
@@ -1,7 +1,13 @@
 # FOSSA CLI Changelog
 
-## v3.9.6
+## v3.9.8
 - Reachability: Users may now provide custom locations for the JAR files emitted by projects and used for reachability analysis ([#1382](https://github.com/fossas/fossa-cli/pull/1382)).
+
+## v3.9.7
+- Add preflight permission checks to validate token type, subscription type, project permissions, and release group permissions [#1383](https://github.com/fossas/fossa-cli/pull/1383)
+
+## v3.9.6
+- Add debug logs for build warnings in `analyze` commands [#1386](https://github.com/fossas/fossa-cli/pull/1386)
 
 ## v3.9.5
 - Maven: Fix hanging maven analysis ([#1381](https://github.com/fossas/fossa-cli/pull/1381)).